+++ packages/console/src/pages/UserDetails/components/UserConnectors.tsx
@@ -113,7 +113,7 @@ const UserConnectors = ({ userId, connectors, onDelete }: Props) => {
                 <tr key={target}>
                   <td>
                     <div className={styles.connectorName}>
-                      <img src={logo} />
+                      <img src={logo} alt="logo" />
                       <div className={styles.name}>
                         <UnnamedTrans resource={name} />
                       </div>