+++ packages/ui/src/components/TermsOfUse/index.tsx
@@ -3,6 +3,7 @@ import { useTranslation } from 'react-i18next';
 
 import Checkbox from '@/components/Checkbox';
 import TextLink from '@/components/TextLink';
+import { onKeyDownHandler } from '@/utils/a11y';
 
 import * as styles from './index.module.scss';
 
@@ -20,12 +21,24 @@ const TermsOfUse = ({ name, className, termsUrl, isChecked, onChange, onTermsCli
 
   const prefix = t('description.agree_with_terms');
 
+  const toggle = () => {
+    onChange(!isChecked);
+  };
+
   return (
     <div
+      role="radio"
+      aria-checked={isChecked}
+      tabIndex={0}
       className={classNames(styles.terms, className)}
-      onClick={() => {
-        onChange(!isChecked);
-      }}
+      onClick={toggle}
+      onKeyDown={onKeyDownHandler({
+        Esc: () => {
+          onChange(false);
+        },
+        Enter: toggle,
+        ' ': toggle,
+      })}
     >
       <Checkbox name={name} checked={isChecked} className={styles.checkBox} />
       <div className={styles.content}>