+++ packages/ui/src/containers/UsernameSignIn/index.tsx
@@ -93,6 +93,7 @@ const UsernameSignIn = ({ className, autoFocus }: Props) => {
     <form className={classNames(styles.form, className)} onSubmit={onSubmitHandler}>
       <div className={styles.formFields}>
         <Input
+          // eslint-disable-next-line jsx-a11y/no-autofocus
           autoFocus={autoFocus}
           className={styles.inputField}
           name="username"